Broke out invalidUrlMessage function to be imported to main.js. #2
